requirements.txt
yt_dlp
openai-whisper
openai